introducing global parameters and parfile

+++ include/physics/kernels.h
@@ -43,7 +43,7 @@ namespace kernels{
   double vector_size(
     point_t& p){
     double result = 0.;
-    for(int i = 0 ; i < gdimension; ++i){
+    for(size_t i = 0 ; i < gdimension; ++i){
       result += p[i]*p[i];
     }
     return sqrt(result);
